requirements.txt
gradio==4.44.0
python-dotenv==1.0.1
pymupdf==1.24.10
openai==1.52.2  # Add OpenAI API library
qdrant-client==1.12.0
neo4j==5.24.0
scikit-learn==1.5.2